requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pytest>=7.4.0
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0 